requirements.txt
streamlit
faiss-cpu
sentence-transformers
beautifulsoup4
requests
PyPDF2
